requirements.txt
numpy==1.26.4
pandas==2.2.2
PyTDC==0.4.17
pytorch_lightning==2.1.3
rdkit==2023.9.6
torch==2.0.1
torch_geometric==2.3.0
tqdm==4.65.0
transformers==4.41.2
tensorboard==2.15.1
mlflow==2.9.2
